+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-...
+- Switch from `babel` to `recast` for JSX decorator.
 
 ## [0.1.0] - 2025-05-05